first stab at sync code; includes snapshots

+++ game/src/main/java/com/snailscuffle/game/accounts/AccountsSnapshot.java
@@ -3,10 +3,10 @@ package com.snailscuffle.game.accounts;
 public class AccountsSnapshot {
 	
 	public final String name;
-	public final long height;
+	public final int height;
 	public final long blockId;
 	
-	public AccountsSnapshot(String name, long height, long blockId) {
+	public AccountsSnapshot(String name, int height, long blockId) {
 		this.name = name;
 		this.height = height;
 		this.blockId = blockId;
